Move debounce factory to misc module
