+++ examples/permalink.html
@@ -17,7 +17,7 @@
   <body ng-controller="MainController as ctrl">
     <app-map app-map="::ctrl.map"></app-map>
     <app-draw app-draw-map="::ctrl.map" app-draw-layer="::ctrl.vectorLayer"></app-draw>
-    <p id="desc">This example shows how to create an application-specific map directive based on the <code><a href="../apidoc/ngeo.map.directive.html" title="Read our documentation">ngeoMap</a></code> directive that uses the <code><a href="../apidoc/ngeo.statemanager.Location.html" title="Read our documentation">ngeoLocation</a></code> and <code><a href="../apidoc/ngeo.Debounce.html" title="Read our documentation">ngeoDebounce</a></code> services to update the browser address bar URL when the map view changes.</p>
+    <p id="desc">This example shows how to create an application-specific map directive based on the <code><a href="../apidoc/ngeo.map.directive.html" title="Read our documentation">ngeoMap</a></code> directive that uses the <code><a href="../apidoc/ngeo.statemanager.Location.html" title="Read our documentation">ngeoLocation</a></code> and <code><a href="../apidoc/ngeo.misc.debounce.html" title="Read our documentation">ngeoDebounce</a></code> services to update the browser address bar URL when the map view changes.</p>
     <script src="../node_modules/jquery/dist/jquery.js"></script>
     <script src="../node_modules/angular/angular.js"></script>
     <script src="../node_modules/angular-gettext/dist/angular-gettext.js"></script>
