+++ options/ngeox.js
@@ -274,6 +274,14 @@ ngeox.MapFeatureOverlayGroup;
 
 
 /**
+ * Provides a debounce function used to debounce calls to a user-provided
+ * function.
+ * @typedef {function(function(?), number, boolean):function()}
+ */
+ngeox.miscDebounce;
+
+
+/**
  * A hash that contains 2 lists of queryable data sources: `wfs` and `wms`.
  * The same data source can only be in one of the two lists. The `wfs` list
  * has priority, i.e. if the data source supports WFS, it's put in the
