+++ src/misc/extraModule.js
@@ -2,6 +2,7 @@ goog.provide('ngeo.misc.extraModule');
 
 goog.require('ngeo');
 goog.require('ngeo.misc.AutoProjection');
+goog.require('ngeo.misc.debounce');
 goog.require('ngeo.misc.EventHelper');
 goog.require('ngeo.misc.FeatureHelper');
 goog.require('ngeo.misc.Time');
@@ -13,6 +14,7 @@ goog.require('ngeo.misc.WMSTime');
 ngeo.misc.extaModule = angular.module('ngeoMiscExtraModule', [
   ngeo.module.name, // Change me when all dependencies are in a module.
   ngeo.misc.AutoProjection.module.name,
+  ngeo.misc.debounce.name,
   ngeo.misc.EventHelper.module.name,
   ngeo.misc.FeatureHelper.module.name,
   ngeo.misc.Time.module.name,
